Add action modal| fix errors
Created modal in the "Call to action" section. Fixed some errors in html/css

Added blue modal box with padding, contrast on text, clickable button-link. Found issue with cascading bleeding over, due to missing Div in html.
Added the div, changed the header links to be valid format, and changed temporary inline picture sizes to be valid as well.
On css side, I renamed a couple of classes and restyled it slightly, to fix any errors the missing div caused.